amazonka-core: Move logging into `Amazonka.Logger`
